Create English to French
Steps of Completion  

Write a Dictionary that converts English numbers to French, one through five. (That would be un, deux, trois, quatre, cinq.) Ask for user input for an English number, and output the French one. Print a nice error message if the text entered isn't in the dictionary. Hint, there is am example in the book that is almost like this. Another hint, this should take 6 lines of code.