Create Pets
Build a list of tuples, where each tuple has info about one pet - type, name, and color. For example,

    animals = [('dog','Fort','brown'),('cat','Tiger','tabby')]
This is a list (notice it is in square brackets) containing 2 items. Each item is a tuple (notice the parentheses) containing that pet's information. Put at least 5 animals (of your choice) in your list. Loop through your list and print it out in this format:

Fort is a brown dog Tiger is a tabby cat

EXTRA CREDIT (5 points each)

1. Also, prompt me to enter one of those colors, and print out all of the pets that are that color.

2.Finally, prompt me to enter one of the pet types, and print out all of the pets of that type.